requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alpaca-trade-api==3.2.0
altair==6.0.0
anaconda-anon-usage @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_835778cmg1/croot/anaconda-anon-usage_1743190150273/work
annotated-types @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/annotated-types_1728588668736/work
anyio==4.9.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==25.3.0
backtrader==1.9.78.123
beautifulsoup4==4.13.4
bitarray==3.8.0
blinker==1.9.0
boltons @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_0a8f8rf2dq/croot/boltons_1737061714423/work
Brotli @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_f7i0oxypt6/croot/brotli-split_1736182464088/work
cachetools==6.2.6
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_50bk79gkw_/croot/certifi_1745939223061/work/certifi
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_e4xd9yd9i2/croot/cffi_1736182819442/work
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
choreographer==1.0.10
ckzg==2.1.5
click==8.2.1
conda @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_7849773gii/croot/conda_1748271012302/work/conda-src
conda-anaconda-telemetry @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_9cczaegl9x/croot/conda-anaconda-telemetry_1744662554568/work
conda-anaconda-tos @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_08c_a7hm7s/croot/conda-anaconda-tos_1744823880434/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1728603284288/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_82jv6q9ech/croot/conda-package-handling_1731369032774/work
conda_package_streaming @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2el5nlypud/croot/conda-package-streaming_1731366187356/work
contourpy==1.3.2
cryptography @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_fdxcx92j10/croot/cryptography_1740577852661/work
curl_cffi==0.11.3
cycler==0.12.1
cytoolz==1.1.0
dash==3.1.1
deprecation==2.1.0
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1728591855492/work
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
fastapi==0.115.14
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
frozendict @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/frozendict_1728605521897/work
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hf-xet==1.1.4
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
hyperframe==6.1.0
idna @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/idna_1728585920859/work
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jsonpatch_1728594556110/work
jsonpointer==2.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.0.0
kalshi-python==2.1.4
kiwisolver==1.4.8
lazy_imports==1.2.0
libmambapy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8fxbc14zt9/croot/mamba-split_1734469521691/work/libmambapy
logistro==1.1.0
markdown-it-py @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/markdown-it-py_1728586321899/work
markdownify==0.13.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/mdurl_1728585992209/work
menuinst @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_66is8p984e/croot/menuinst_1738943433761/work
mpmath==1.3.0
msgpack==1.0.3
multidict==6.5.0
multitasking==0.0.11
mysql-connector-python==9.3.0
narwhals==1.45.0
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.3
orjson==3.11.3
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a6_qk3qyg7/croot/packaging_1734472142254/work
pandas==2.3.0
parsimonious==0.10.0
patsy==1.0.1
peewee==3.18.1
pillow==11.2.1
platformdirs @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4euggz8v1n/croot/platformdirs_1744273055841/work
plotly==6.2.0
pluggy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_bei6q1cvvt/croot/pluggy_1733169621717/work
poly_eip712_structs==0.0.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.5
py_order_utils==0.3.2
pyaml==25.5.0
pyarrow==22.0.0
pycosat @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_61k1o_oje7/croot/pycosat_1736868429741/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.23.0
pydantic @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4blmh3bau_/croot/pydantic_1734736077970/work
pydantic_core @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_bdqco4zmod/croot/pydantic-core_1734726070262/work
pydeck==0.9.1
Pygments @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_f0f10r98sf/croot/pygments_1744664126614/work
pyparsing==3.2.3
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1728586068722/work
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jobspy==1.1.82
python-telegram-bot==22.1
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.1
redis==6.4.0
referencing==0.37.0
regex==2024.11.6
reportlab==4.4.1
requests @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_ee45nsd33z/croot/requests_1730999134038/work
retrying==1.4.0
rich @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_034myiu7pr/croot/rich_1732638981241/work
rlp==4.1.0
rpds-py==0.30.0
ruamel.yaml @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_beoui95smx/croot/ruamel.yaml_1745960312720/work
ruamel.yaml.clib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d9xusw0yil/croot/ruamel.yaml.clib_1745937162586/work
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.7.0
scikit-optimize==0.10.2
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
statsmodels==0.14.5
streamlit==1.53.1
sympy==1.14.0
ta==0.11.0
tenacity==9.1.2
threadpoolctl==3.6.0
tls-client==1.0.1
tokenizers==0.21.1
toml==0.10.2
toolz==1.1.0
torch==2.7.1
tornado==6.5.4
tqdm @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5642tzrprr/croot/tqdm_1738943467095/work
transformers==4.52.4
truststore @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a1qa3ogm4o/croot/truststore_1736550138575/work
types-requests==2.32.4.20260107
typing_extensions @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_0b3jpv_f79/croot/typing_extensions_1734714864260/work
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.35.0
watchdog==6.0.0
web3==7.14.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wheel==0.45.1
xgboost==3.0.2
yarl==1.20.1
yfinance==0.2.64
zipp==3.23.0
zstandard @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_a5_i6g4o6n/croot/zstandard_1731356352787/work
